+++ website/src/components/index.ts
@@ -2,3 +2,4 @@ export { Button, IconButton, CleanButton } from "./button"
 export { default as Link } from "./link"
 export { Select, Label } from "./select"
 export { AudioPlayer } from "./audio-player"
+export { Breadcrumbs } from "./breadcrumbs"
