+++ website/src/pages/documents/document.page.tsx
@@ -6,8 +6,7 @@ import { Helmet } from "react-helmet"
 import { MdSettings } from "react-icons/md"
 import { Dialog, DialogBackdrop, useDialogState } from "reakit/Dialog"
 import { Tab, TabList, TabPanel } from "reakit/Tab"
-import { Breadcrumbs } from "src/breadcrumbs"
-import { AudioPlayer, Button, Link } from "src/components"
+import { AudioPlayer, Breadcrumbs, Button, Link } from "src/components"
 import { useMediaQuery } from "src/custom-hooks"
 import * as Dailp from "src/graphql/dailp"
 import Layout from "src/layout"
